Merges toggle.cpp/h into common from redalert.
