+++ common/toggle.h
